Fix for setting the proton charge. It's stored as a float in the Nexus file and trying to retrieve it as a double gives you zero. The unit test wasn't testing this as I thought it was. Updated it to do so. Re #1416

+++ Code/Mantid/Nexus/src/LoadISISNexus.cpp
@@ -597,7 +597,7 @@ namespace Mantid
       getNexusData(&rpb_int[0]);
       closeNexusData();
 
-      double rpb_dbl[32];
+      float rpb_dbl[32];
       openNexusData("RRPB");
       getNexusData(&rpb_dbl[0]);
       closeNexusData();
